Lazy-update status bar
The StatusBar should not be updated every time a prop changes, it
should be updated when one of the visual props changes: `style`,
`hidden` or `tintColor`.

The current aggressive update strategy causes problems in apps with
multiple nav bars, since bars on views hidden lower in the route stack
can end up operating on the status bar, which is obviously unwanted
behavior.